02-Using-gem5/03-running-in-gem5 Marp slides with images
This commit adds Marp slides at `slides/02-Using-gem5/03-running-in-gem5.md`, ported
from `gem5/bootcamp 2022/slides-archive/Using gem5/03-gem5-running-tutorial` in the
OneDrive. It also adds the images used in the Marp slides. It also changes the section
text color in `gem5.css` to be black, as the text is white otherwise.