03/03: Changed slide spacing, added links, added commands
This commit changes the slide spacing and adds more slides
as necessary so it is easier to read, adds links to the
completed examples where appropriate, and removes previously
added links that point to files that the audience will work on.

It also adds commands to allow the audience to move the completed
files to the gem5 directory to run them.